oura2bq_gcf/requirements.txt
pandas
google-cloud-secret-manager
oura
google-cloud-bigquery
